pothole-backend/requirements.txt
flask
flask-cors
pymongo
ultralytics
python-dotenv
geopy
PyJWT
werkzeug